subcommand traits
